+++ ubi/src/lib.rs
@@ -32,8 +32,9 @@
 //! with the same name as the project. In either case, the file will be installed with the name it
 //! has in the archive file.
 //!
-//! If the release is in the form of a bare executable or a compressed executable, then the installed
-//! executable will use the name of the project instead.
+//! If the release is in the form of a bare executable or a compressed executable, then the
+//! installed executable will use the name of the project instead. For files with a `.exe`, `.pyz`
+//! or `.AppImage`, the installed executable will be `$project_name.$extension`.
 //!
 //! This is a bit inconsistent, but it's how `ubi` has behaved since it was created, and I find this
 //! to be the sanest behavior. Some projects, for example `rust-analyzer`, provide releases as